Change encounter slot probabilities
This is an attempt to reduce tedium by eliminating 1% encounter slots, which should also free up encounter slot space on routes that were doubling up on the 4% and 1% slots for the same mon, allowing realistic encounter chances for a wider variety of mons per route.

TODO: Adjust actual wild routes so that this isn't completely chaotic.